Add Effect dsp
Some of the Effect parameters change assignment and ranges depending
on the Effect type. This adds complexity to the Parameter model which
which was mainly designed on const range dicrete parameter ranges.

This commit adds more flexibility to the DiscreteParameter model and
introduces variable range discrete parameters. The range for these
parameters is controller by a discriminant which is derived from
a constant range parameter. E.g. for Effect, the effect type is
the discriminant for Speed, Depth and Regen. All three being
variable range parameters. Regen introduces additional variation
as it is only active for a subset of the effect types.