Implement Normal using an u8
Normal used to be implemented with an f32 which eased computation
and provided good precision.

This commit uses an `u8` instead in order to reduce the size of
the dicrete parameters and DSPs, thus improving cache locality
and speeding up `struct` copies.

Internal computation uses an `u32` to reduce precision loss.