Use RawValue as the inner type for discrete Parameters
In a previous commit, discrete Parameters inner type was changed
from f32 to a Normal based on u8. This reduces the size of the dsps.
In retrospect, using a Normal as a base for the Parameters isn't
ideal: it doesn't match any other representations and requires
conversions for any interactions.

This commit keeps the u8 inner type, but uses the RawValue, which is
the value returned by the device in program procedures.