Add Programs Bank support
The new pick list gives access to user and factory Programs Banks.
User programs are automatically downloaded from the device when
the J-Station is identified. Factory programs are downloaded on
demand, either after a user selection from the programs list in
Factory mode or when rotating the data button on the device (the
factory programs appear with a dot on the digital display).

Note that this feature was not implemented in gstation-edit.