Implement Store and improve initial prog identification
Implement the Store sequence.

Previous program identification compared the name and raw buffers
between the program received as part of a ProgramUpdateResp and
the program list received as OneProgramResps. Unfortunately, the
raw buffers can differ with regard to the inactive parameters
(e.g. Regen). This problem was hidden in gstation-edit because
the first program was always selected in case the identification
failed.