Remove unnecessary unsafe get_unchecked
Parameter access to the Program data array where using get_unchecked
so as to avoid potential bound checks. This is unnecessary as the
array is fixed sized and the index (ProgramNumber) are const and
are guaranteed to be in the range of the array. This should ensure
bound checks to be optimized out by the compiler.