jstation: impl the ProcedureId trait
This `trait` defines the Id and Version of a `struct` which can be built
or / and parsed as a `Procedure`. Previously the Id and Version were
either part of the `ProcedureBuilder` impl or as `const`s on the `struct`
`impl` if no `ProcedureBuilder` was defined. The new `trait` provides a
single way to identify procedures.